+++ CHANGELOG.md
@@ -19,6 +19,7 @@ Please view this file on the master branch, on stable branches it's out of date.
 - Anonymous user can enter service offer selection section (@mkasztelnik)
 - Service access button (@martaswiatkowska)
 - Services can be added to project multiple times if offer has parameters (@martaswiatkowska)
+- Service and offer type enum values (@goreck888)
 
 ### Deprecated